Refactored serialization module
